feat: introduce page issue feed
that presents the user with issues of others

+++ support-pipe-quarkus/src/main/java/de/simonlammes/issue/Issue.java
@@ -12,13 +12,13 @@ public class Issue {
     private String title;
     private String description;
     @Column(name = "creator_id")
-    private int creatorId;
+    private long creatorId;
 
-    public int getCreatorId() {
+    public long getCreatorId() {
         return creatorId;
     }
 
-    public void setCreatorId(int creatorId) {
+    public void setCreatorId(long creatorId) {
         this.creatorId = creatorId;
     }
